services/screen_capture/requirements.txt
mcp[cli]>=1.4.1
Pillow>=10.0.0
pyautogui>=0.9.54